#163 modify use_case_factory
・add next_holiday
・call common module method at each class.
・refactor dates_drive_builder
