+++ lib/holidays/use_case_factory.rb
@@ -1,4 +1,6 @@
+require 'holidays/use_case/context/context_common'
 require 'holidays/use_case/context/between'
+require 'holidays/use_case/context/next_holiday'
 require 'holidays/use_case/context/dates_driver_builder'
 
 module Holidays
@@ -6,6 +8,14 @@ module Holidays
     class << self
       def between
         UseCase::Context::Between.new(
+          DefinitionFactory.holidays_by_month_repository,
+          DateCalculatorFactory.day_of_month_calculator,
+          DefinitionFactory.custom_methods_repository,
+          DefinitionFactory.proc_result_cache_repository,
+        )
+      end
+      def next_holiday
+        UseCase::Context::NextHoliday.new(
           DefinitionFactory.holidays_by_month_repository,
           DateCalculatorFactory.day_of_month_calculator,
           DefinitionFactory.custom_methods_repository,
